+++ NEWS.md
@@ -6,7 +6,7 @@
   those exclusively affiliated with the species' passage distributions and (ii)
   habitat types for resident distributions of non-migratory species would not
   include those exclusively affiliated with the species' passage distributions.
-  Thanks to Jianqiao Zhao for raising this issue.
+  Thanks to Jianqiao Zhao for bug report.
 - Update built-in testing data to include multiple migratory species.
 - Update internal R script for creating test dataset
   (i.e., `inst/scripts/test-data.R`) to be compatible with current version of